🐛 BigDecimal을 사용하지 못하는 문제 해결

+++ java-2th-weekly/Beauty.java
@@ -1,7 +1,7 @@
 import java.math.BigDecimal;
 
 public class Beauty extends Product{
-    Beauty(String name, BigDecimal price, Double weight){
+    Beauty(String name, int price,double weight){
         super(name, price, weight);
     }
 }
