+++ java-2th-weekly/Grocery.java
@@ -1,7 +1,7 @@
 import java.math.BigDecimal;
 
 public class Grocery extends Product{
-    Grocery(String name, BigDecimal price, Double weight){
+    Grocery(String name, int price, double weight){
         super(name, price, weight);
     }
 }
