+++ java-2th-weekly/LargeAppliance.java
@@ -1,7 +1,7 @@
 import java.math.BigDecimal;
 
 public class LargeAppliance extends Product{
-    LargeAppliance(String name, BigDecimal price, Double weight){
+    LargeAppliance(String name, int price, double weight){
         super(name, price, weight);
     }
 }
